refactor: add index files with multiple imports
