+++ src/main/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Login from '@presentation/pages/login'
+import { Login } from '@presentation/pages'
 
 ReactDOM.render(<Login />, document.getElementById('main'))
