+++ src/presentation/components/index.ts
@@ -1,0 +1,6 @@
+export { default as Footer } from './footer/footer'
+export { default as FormStatus } from './form-status/form-status'
+export { default as Input } from './input/input'
+export { default as Header } from './login-header/login-header'
+export { default as Logo } from './logo/logo'
+export { default as Sprinner } from './spinner/spinner'
