+++ src/presentation/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as Login } from './login/login'
