+++ src/presentation/pages/login/login.tsx
@@ -1,8 +1,5 @@
-import Footer from '@presentation/components/footer/footer'
-import FormStatus from '@presentation/components/form-status/form-status'
-import Input from '@presentation/components/input/input'
-import Header from '@presentation/components/login-header/login-header'
 import React from 'react'
+import { Input, FormStatus, Footer, Header } from '@presentation/components'
 import Styles from './login-styles.scss'
 
 const Login: React.FC = () => {
